refactor: 칸반보드 관련 component들 세부 폴더로 옮김
