+++ app/(main)/_components/KanbanBoard/board-trash-box.tsx
